requirements.txt
colorlog==6.7.0
easydict==1.10
matplotlib==3.5.3
numpy==1.21.5
opencv_contrib_python==4.6.0.66
opencv_python==4.7.0.68
Pillow==10.0.0
PyYAML==6.0
spikingjelly==0.0.0.0.6
timm==0.5.4
torch==1.8.0
torchvision==0.9.0
tqdm==4.64.1
